% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinadevfin2-package.R
\docType{package}
\name{chinadevfin2-package}
\alias{chinadevfin2}
\alias{chinadevfin2-package}
\title{chinadevfin2: An R Package for AidData's Global Chinese Development Finance 2.0 Dataset}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

`chinadevfin2` is a package for working with AidData's Global Chinese Development Finance 2.0 Dataset. The dataset includes 13,427 projects worth $843 billion across 165 countries. `chinadevfin2` provides an easy way for R users to load the dataset, and helper functions to ease common analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/t-emery/chinadevfin2}
  \item \url{https://t-emery.github.io/chinadevfin2/}
  \item Report bugs at \url{https://github.com/t-emery/chinadevfin2/issues}
}

}
\author{
\strong{Maintainer}: Teal Emery \email{lte@tealinsights.com}

}
\keyword{internal}
