% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdf2_data_dictionary.R
\docType{data}
\name{gcdf2_data_dictionary}
\alias{gcdf2_data_dictionary}
\title{Data Dictionary for the AidData Global Chinese Development Finance Dataset}
\format{
A tibble with 70 rows and 2 variables:
\describe{
\item{field_name}{chr The name of the variable.}
\item{description}{chr The variable's definition.}

}
}
\source{
\url{https://www.aiddata.org/data/aiddatas-global-chinese-development-finance-dataset-version-2-0}
}
\usage{
gcdf2_data_dictionary
}
\description{
A data dictionary for the \code{gcdf2_dataset}
}
\keyword{datasets}
