% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdf2_dataset.R
\docType{data}
\name{gcdf2_dataset}
\alias{gcdf2_dataset}
\title{AidData Global Chinese Development Finance Dataset, Version 2.0}
\format{
A tibble with 13,427 rows and 70 variables:
\describe{
\item{Dataset}{\code{gcdf2_data_dictionary} contains definitions of all 70 variables.}

}
}
\source{
\url{https://www.aiddata.org/data/aiddatas-global-chinese-development-finance-dataset-version-2-0}
}
\usage{
gcdf2_dataset
}
\description{
Dataset compiled by AidData containing 13,427 development projects worth $843 billion financed by more than 300 Chinese government institutions and state-owned entities across 165 countries in every major region of the world from 2000-2017.
}
\keyword{datasets}
