% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{get_gcdf2_dataset}
\alias{get_gcdf2_dataset}
\title{Load the AidData Global Chinese Development Finance 2.0 Dataset}
\usage{
get_gcdf2_dataset(standardized_countrynames = TRUE)
}
\arguments{
\item{standardized_countrynames}{set to \code{TRUE} by default, this argument attaches standardized country names to make it easier to join the dataset with other datasets.}
}
\value{
\code{get_gcdf2_dataset()} returns a tibble with 13,427 observations.
}
\description{
\code{get_gcdf2_dataset()} makes it easy to load \href{https://www.aiddata.org/data/aiddatas-global-chinese-development-finance-dataset-version-2-0}{AidData's Global Chinese Development Finance Dataset, Version 2.0}
(GCDF 2.0). By default, the function adds standardized country names that make it easier to merge it with other datasets.  The standardized country names are described in the documentation for \link{gcdf2_standardized_countrynames}.
Setting \code{standardized_countrynames = FALSE} is equivalent to simply calling the \link{gcdf2_dataset} data object.  The data definitions are available in the \link{gcdf2_data_dictionary}.
}
\examples{
# returns the GCDF 2.0 dataset, with standardized country names
get_gcdf2_dataset()

# returns the GCDF 2.0 dataset as-is.
get_gcdf2_dataset(standardized_countrynames = FALSE)

# the latter is equivalent to simply calling the [gcdf2_dataset] data object
gcdf2_dataset
}
